Insee Zip code Cities of 01 Dpt, France.txt
It's a INSEE zip code attached to the city in France who all cities started with 01.
The two first number "01" are the French Department named : Ain
They are 95 Department in France Metropolitain with 2 numbers and the other abroad are named with 3 numbers.
I am working on the others. (Is going to take time).
By the way, I am working on the street's names two.